\docType{data}
\name{Idgf2}
\alias{Idgf2}
\title{Example Data in Vignette}
\format{A list with 5 arguments:}
\usage{
data(Idgf2)
}
\description{
The data 'Idgf2'(list) has 5 arguments, including 'data', 'exon.length', 'gene.names', 'group.label', 'reads'.
Please refer to 'timeSeq.Rd' for more information.
}
\examples{
data(Idgf2)
str(Idgf2)
}
\keyword{datasets}
