\docType{data}
\name{geneset.fit}
\alias{geneset.fit}
\title{Object returned by timeSeq.geneset in the simualtion example}
\format{A list with 5 arguments:}
\usage{
data(object_by_timeSeq.geneset)
}
\description{
The dataset is generated according to the simulation example.
The data 'geneset.fit'(list) has 5 arguments, including 'NPDE.ratio', 'PDE.ratio', 
'NPDE.pvalue', 'PDE.pvalue', 'geneset.names'.
Please refer to 'timeSeq.geneset.Rd' for more information.
}
\examples{
data(object_by_timeSeq.geneset)
str(geneset.fit)
}
\keyword{datasets}