\docType{data}
\name{model.fit}
\alias{model.fit}
\title{Object returned by timeSeq in the simualtion example}
\format{A list with 17 arguments:}
\usage{
data(object_by_timeSeq)
}
\description{
The dataset is generated according to the simulation example.
The data 'model.fit'(list) has 17 arguments, including 'sorted', 'count', 'NPDE.ratio', 'PDE.ratio', 'genenames',
 'table', 'data', 'gene.names', 'effective.lib.size', 'exon.length', 'group.label', 'group.length',
 'group1.length', 'group2.length', 'exon.level', 'p.values', 'pvalue'.
Please refer to 'timeSeq.Rd' for more information.
}
\examples{
data(object_by_timeSeq)
str(model.fit)
}
\keyword{datasets}
