\docType{data}
\name{pAbp}
\alias{pAbp}
\title{Example Data in Vignette}
\format{A list with 4 arguments:}
\usage{
data(pAbp)
}
\description{
The data 'pAbp'(list) has 4 arguments, including 'data.count', 'exon.length', 'gene.names', and 'group.label'.
Please refer to 'timeSeq.Rd' for more information.
}
\examples{
data(pAbp)
str(pAbp)
}
\keyword{datasets}
