\docType{data}
\name{simulate.dt}
\alias{simulate.dt}
\title{Example Data in Vignette}
\format{A list with 5 arguments:}
\usage{
data(simulate.dt)
}
\description{
The data 'simulate.dt'(list) has 4 arguments, including 'data.count', 'gene.names', 'group.label', 'reads'.
Please refer to 'timeSeq.Rd' for more information.
}
\examples{
data(simulate.dt)
str(simulate.dt)
}
\keyword{datasets}
