\name{timeSeq_1.0.0-package}
\alias{timeSeq_1.0.0-package}
\alias{timeSeq_1.0.0}
\docType{package}
\title{
Statistical Inference for Time Course RNA-Seq Data using a Negative Binomial Mixed-Effects Model
}
\description{
In this article, we propose a negative binomial mixed-effects (NBME) model to identify DE genes, including nonparallel differentially expressed (NPDE) and parallel differentially expressed (PDE) genes, in time course RNA-seq data. A significance test for detecting DE genes is derived using the Kullback-Leibler distance ratio. Additionally, a significance test for gene sets is developed using a gene set score.
}
\details{
\tabular{ll}{
Package: \tab timeSeq\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-4-19\cr
License: \tab GPL-2 | GPL-3\cr
}

}
\author{
Fan Gao and Xiaoxiao Sun

Maintainer: Fan Gao <fangaohz@gmail.com>
}
\references{
Chong Gu. Model diagnostics for smoothing spline ANOVA models. Canadian Journal of Statistics, 32(4):347-358, 2004.

Chong Gu. Smoothing spline ANOVA models. Springer, second edition, 2013.

Chong Gu and Ping Ma. Optimal smoothing in nonparametric mixed-effect models. Annals of Statistics, pages 1357-1379, 2005a.
}
\keyword{ package }
