\name{timeSeq_1.0.4-package}
\alias{timeSeq_1.0.4-package}
\alias{timeSeq_1.0.4}
\docType{package}
\title{
Statistical Inference for Time Course RNA-Seq Data using a Negative Binomial Mixed-Effects Model
}
\description{
In this package, we propose a negative binomial mixed-effects (NBME) model to identify differentially expressed (DE) genes, including nonparallel differentially expressed (NPDE) and parallel differentially expressed (PDE) genes, in time course RNA-seq data. 
}
\details{
\tabular{ll}{
Package: \tab timeSeq\cr
Type: \tab Package\cr
Version: \tab 1.0.4\cr
Date: \tab 2019-2-08\cr
License: \tab GPL-2 | GPL-3\cr
}

}
\author{
Fan Gao and Xiaoxiao Sun

Maintainer: Fan Gao <fangaohz@gmail.com>
}
\references{
Sun, Xiaoxiao, David Dalpiaz, Di Wu, Jun S. Liu, Wenxuan Zhong, and Ping Ma. "Statistical inference for time course RNA-Seq data using a negative binomial mixed-effect model." BMC Bioinformatics, 17(1):324, 2016.

Chong Gu. Model diagnostics for smoothing spline ANOVA models. Canadian Journal of Statistics, 32(4):347-358, 2004.

Chong Gu. Smoothing spline ANOVA models. Springer, second edition, 2013.

Chong Gu and Ping Ma. Optimal smoothing in nonparametric mixed-effect models. Annals of Statistics, pages 1357-1379, 2005a.

Wood (2001) mgcv:GAMs and Generalized Ridge Regression for R. R News 1(2):20-25
}
\keyword{ package }
