\docType{data}
\name{geneset.fit}
\alias{geneset.fit}
\title{Object returned by timeSeq.geneset for simualtion examples}
\format{A list with 5 arguments:}
\usage{
data(object_by_timeSeq.geneset)
}
\description{
The data are generated according to the simulation example.
The data 'geneset.fit'(list) has 5 arguments, including 'NPDE.ratio', 'PDE.ratio', 
'NPDE.pvalue', 'PDE.pvalue', 'geneset.names'.
You could learn their meanings by referring to 'timeSeq.geneset.Rd'.
}
\examples{
data(object_by_timeSeq.geneset)
str(geneset.fit)
}
\keyword{datasets}